prepare-root: Factor out composefs handling into otcore
I'm thinking about creating ostree-prepare-soft-reboot.c.
Prepare for this by factoring out shared helper functions.
