+++ src/libotcore/otcore.h
@@ -78,6 +78,27 @@ G_DEFINE_AUTOPTR_CLEANUP_FUNC (ComposefsConfig, otcore_free_composefs_config)
 ComposefsConfig *otcore_load_composefs_config (const char *cmdline, GKeyFile *config,
                                                gboolean load_keys, GError **error);
 
+/**
+ * otcore_mount_composefs:
+ * @composefs_config: Configuration for composefs.
+ * @metadata_builder: (transfer none): GVariantBuilder to add metadata to.
+ * @root_transient: Whether the root filesystem is transient.
+ * @root_mountpoint: The mount point of the physical root filesystem.
+ * @deploy_path: The path to the deployment.
+ * @mount_target: The target path to mount the composefs image.
+ * @out_using_composefs: (out): Whether composefs was successfully used.
+ * @error: (out): Return location for a GError, or %NULL.
+ *
+ * Mounts a composefs image based on the provided configuration.
+ *
+ * Returns: %TRUE on success, %FALSE on error.
+ */
+gboolean otcore_mount_composefs (ComposefsConfig *composefs_config,
+                                 GVariantBuilder *metadata_builder, gboolean root_transient,
+                                 const char *root_mountpoint, const char *deploy_path,
+                                 const char *mount_target, bool *out_using_composefs,
+                                 GError **error);
+
 // Our directory with transient state (eventually /run/ostree-booted should be a link to
 // /run/ostree/booted)
 #define OTCORE_RUN_OSTREE "/run/ostree"
